+++ builder/src/main/java/com/stratio/cassandra/lucene/builder/index/schema/mapping/GeoPointMapper.java
@@ -40,6 +40,14 @@ public class GeoPointMapper extends Mapper<GeoPointMapper> {
     @JsonProperty("max_levels")
     Integer maxLevels;
 
+    /** The name of the column containing the longitude. */
+    @JsonProperty("indexed")
+    Boolean indexed;
+
+    /** The maximum number of levels in the tree. */
+    @JsonProperty("sorted")
+    Boolean sorted;
+
     /**
      * Builds a new {@code GeoPointMapper}.
      *
@@ -61,6 +69,28 @@ public class GeoPointMapper extends Mapper<GeoPointMapper> {
      */
     public GeoPointMapper maxLevels(Integer maxLevels) {
         this.maxLevels = maxLevels;
+        return this;
+    }
+
+    /**
+     * Sets if the field supports searching.
+     *
+     * @param indexed if the field supports searching.
+     * @return This.
+     */
+    public final GeoPointMapper indexed(Boolean indexed) {
+        this.indexed = indexed;
+        return this;
+    }
+
+    /**
+     * Sets if the field supports sorting.
+     *
+     * @param sorted if the field supports sorting.
+     * @return This.
+     */
+    public final GeoPointMapper sorted(Boolean sorted) {
+        this.sorted = sorted;
         return this;
     }
 }