+++ plugin/src/main/java/com/stratio/cassandra/lucene/schema/mapping/GeoPointMapper.java
@@ -24,9 +24,18 @@ import com.spatial4j.core.shape.Point;
 import com.stratio.cassandra.lucene.IndexException;
 import com.stratio.cassandra.lucene.schema.column.Column;
 import com.stratio.cassandra.lucene.schema.column.Columns;
-import org.apache.cassandra.db.marshal.*;
+import org.apache.cassandra.db.marshal.AsciiType;
+import org.apache.cassandra.db.marshal.DecimalType;
+import org.apache.cassandra.db.marshal.DoubleType;
+import org.apache.cassandra.db.marshal.FloatType;
+import org.apache.cassandra.db.marshal.Int32Type;
+import org.apache.cassandra.db.marshal.IntegerType;
+import org.apache.cassandra.db.marshal.LongType;
+import org.apache.cassandra.db.marshal.ShortType;
+import org.apache.cassandra.db.marshal.UTF8Type;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.lucene.document.Document;
+import org.apache.lucene.document.StoredField;
 import org.apache.lucene.index.IndexableField;
 import org.apache.lucene.search.SortField;
 import org.apache.lucene.spatial.SpatialStrategy;
@@ -36,6 +45,7 @@ import org.apache.lucene.spatial.prefix.tree.GeohashPrefixTree;
 import org.apache.lucene.spatial.prefix.tree.SpatialPrefixTree;
 
 import java.util.Arrays;
+
 
 /**
  * A {@link Mapper} to map geographical points.
@@ -52,10 +62,10 @@ public class GeoPointMapper extends Mapper {
     public static final SpatialContext SPATIAL_CONTEXT = SpatialContext.GEO;
     public static final int DEFAULT_MAX_LEVELS = 11;
 
-    /** Tha name of the latitude column. */
+    /** The name of the latitude column. */
     public final String latitude;
 
-    /** Tha name of the longitude column. */
+    /** The name of the longitude column. */
     public final String longitude;
 
     /** The max number of levels in the tree. */
@@ -76,10 +86,10 @@ public class GeoPointMapper extends Mapper {
      * @param longitude The name of the column containing the longitude.
      * @param maxLevels The maximum number of levels in the tree.
      */
-    public GeoPointMapper(String field, Boolean validated, String latitude, String longitude, Integer maxLevels) {
+    public GeoPointMapper(String field, Boolean indexed, Boolean sorted, Boolean validated, String latitude, String longitude, Integer maxLevels) {
         super(field,
-              true,
-              false,
+              indexed,
+              sorted,
               validated,
               null,
               Arrays.asList(latitude, longitude),
@@ -121,7 +131,7 @@ public class GeoPointMapper extends Mapper {
         if (latitude == null) {
             throw new IndexException("%s required", name);
         } else if (latitude < MIN_LATITUDE || latitude > MAX_LATITUDE) {
-            throw new IndexException("%s must be in range [%s, %s], but found '%s'",
+            throw new IndexException("%s must be in range [%s, %s], but found %s",
                                      name,
                                      MIN_LATITUDE,
                                      MAX_LATITUDE,
@@ -167,18 +177,23 @@ public class GeoPointMapper extends Mapper {
 
         Point point = SPATIAL_CONTEXT.makePoint(lon, lat);
 
-        for (IndexableField field : distanceStrategy.createIndexableFields(point)) {
-            document.add(field);
+        if (indexed) {
+            for (IndexableField field : distanceStrategy.createIndexableFields(point)) {
+                document.add(field);
+            }
+            for (IndexableField field : bboxStrategy.createIndexableFields(point)) {
+                document.add(field);
+            }
         }
-        for (IndexableField field : bboxStrategy.createIndexableFields(point)) {
-            document.add(field);
+        if (sorted) {
+            document.add(new StoredField(distanceStrategy.getFieldName(), point.getX()+" "+point.getY()));
         }
     }
 
     /** {@inheritDoc} */
     @Override
     public SortField sortField(String name, boolean reverse) {
-        throw new IndexException("GeoPoint mapper '%s' does not support sorting", name);
+        throw new IndexException("GeoPoint mapper '%s' does not support simple sorting", name);
     }
 
     /**
@@ -186,8 +201,8 @@ public class GeoPointMapper extends Mapper {
      *
      * @param columns The {@link Columns} containing the latitude.
      */
-    Double readLatitude(Columns columns) {
-        Column<?> column = columns.getColumnsByCellName(latitude).getFirst();
+    public Double readLatitude(Columns columns) {
+        Column<?> column = columns.getColumnsByFullName(latitude).getFirst();
         return column == null ? null : readLatitude(column.getComposedValue());
     }
 
@@ -197,8 +212,8 @@ public class GeoPointMapper extends Mapper {
      *
      * @param columns The {@link Columns} containing the latitude.
      */
-    Double readLongitude(Columns columns) {
-        Column<?> column = columns.getColumnsByCellName(longitude).getFirst();
+    public Double readLongitude(Columns columns) {
+        Column<?> column = columns.getColumnsByFullName(longitude).getFirst();
         return column == null ? null : readLongitude(column.getComposedValue());
     }
 
@@ -223,6 +238,16 @@ public class GeoPointMapper extends Mapper {
         }
         return checkLatitude("latitude", value);
     }
+
+    /**
+     * Returns the distance {@link SpatialStrategy}.
+     *
+     * @return The {@link SpatialStrategy}.
+     */
+    public SpatialStrategy getDistanceStrategy() {
+        return distanceStrategy;
+    }
+
 
     /**
      * Returns the longitude contained in the specified {@link Object}.
@@ -257,4 +282,7 @@ public class GeoPointMapper extends Mapper {
                       .add("maxLevels", maxLevels)
                       .toString();
     }
+
+
+
 }