+++ plugin/src/test/java/com/stratio/cassandra/lucene/search/sort/builder/SortBuilderTest.java
@@ -25,7 +25,9 @@ import com.stratio.cassandra.lucene.util.JsonSerializer;
 import org.junit.Test;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.List;
 
 import static org.junit.Assert.*;
 
@@ -38,8 +40,8 @@ public class SortBuilderTest {
 
     @Test
     public void testBuildWithArray() {
-        SortFieldBuilder sortFieldBuilder1 = new SortFieldBuilder("field1").reverse(true);
-        SortFieldBuilder sortFieldBuilder2 = new SortFieldBuilder("field2").reverse(false);
+        SimpleSortFieldBuilder sortFieldBuilder1 = new SimpleSortFieldBuilder("field1").reverse(true);
+        GeoDistanceSortFieldBuilder sortFieldBuilder2 =  new GeoDistanceSortFieldBuilder("field2",0.0,0.0).reverse(true);
         SortBuilder sortBuilder = new SortBuilder(sortFieldBuilder1, sortFieldBuilder2);
         Sort sort = sortBuilder.build();
         assertNotNull("Sort is not built", sort);
@@ -53,9 +55,13 @@ public class SortBuilderTest {
 
     @Test
     public void testBuildWithList() {
-        SortFieldBuilder sortFieldBuilder1 = new SortFieldBuilder("field1").reverse(true);
-        SortFieldBuilder sortFieldBuilder2 = new SortFieldBuilder("field2").reverse(false);
-        SortBuilder sortBuilder = new SortBuilder(Arrays.asList(sortFieldBuilder1, sortFieldBuilder2));
+        SimpleSortFieldBuilder sortFieldBuilder1 = new SimpleSortFieldBuilder("field1").reverse(true);
+        GeoDistanceSortFieldBuilder sortFieldBuilder2 = new GeoDistanceSortFieldBuilder("field2",0.0,0.0).reverse(true);
+        List<SortFieldBuilder> sortFieldBuilderList= new ArrayList<>();
+        sortFieldBuilderList.add(sortFieldBuilder1);
+        sortFieldBuilderList.add(sortFieldBuilder2);
+        SortBuilder sortBuilder = new SortBuilder(sortFieldBuilderList);
+
         Sort sort = sortBuilder.build();
         assertNotNull("Sort is not built", sort);
         assertArrayEquals("List based builder is wrong",
@@ -65,15 +71,29 @@ public class SortBuilderTest {
 
     @Test
     public void testJson() throws IOException {
-        SortFieldBuilder sortFieldBuilder1 = new SortFieldBuilder("field1").reverse(true);
-        SortFieldBuilder sortFieldBuilder2 = new SortFieldBuilder("field2").reverse(false);
-        SortFieldBuilder sortFieldBuilder3 = new SortFieldBuilder("field3");
+        SimpleSortFieldBuilder sortFieldBuilder1 = new SimpleSortFieldBuilder("field1").reverse(true);
+        GeoDistanceSortFieldBuilder sortFieldBuilder2 = new GeoDistanceSortFieldBuilder("mapper2",0.0,0.0).reverse(true);
+        SimpleSortFieldBuilder sortFieldBuilder3 = new SimpleSortFieldBuilder("field3");
         SortBuilder sortBuilder = new SortBuilder(sortFieldBuilder1, sortFieldBuilder2, sortFieldBuilder3);
         String json = JsonSerializer.toString(sortBuilder);
-        assertEquals("Method #toString is wrong", "{fields:[{field:\"field1\",reverse:true}," +
-                                                  "{field:\"field2\"," +
-                                                  "reverse:false}," +
-                                                  "{field:\"field3\",reverse:false}]}", json);
+        assertEquals("Method #toString is wrong", "{fields:[{type:\"simple\",field:\"field1\",reverse:true}," +
+                                                  "{type:\"geo_distance\",mapper:\"mapper2\",longitude:0.0,latitude:0.0,reverse:true}," +
+                                                  "{type:\"simple\",field:\"field3\",reverse:false}]}", json);
 
+    }
+
+    @Test
+    public void testDeserializeDefaultSort() {
+        String json1="{field:\"field1\",reverse:true}";
+
+        SortFieldBuilder sortFieldBuilder = null;
+        try {
+            sortFieldBuilder = JsonSerializer.fromString(json1, SortFieldBuilder.class);
+            assertEquals("JSON serialization is wrong",sortFieldBuilder.getClass(),SimpleSortFieldBuilder.class);
+            String json2 = JsonSerializer.toString(sortFieldBuilder);
+            assertEquals("JSON serialization is wrong", "{type:\"simple\",field:\"field1\",reverse:true}", json2);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
     }
 }